Update audio transcription system
test

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/AEGISL0L/audio_stuff?shareId=XXXX-XXXX-XXXX-XXXX).
