+++ audio_project/requirements.txt.txt
@@ -1,0 +1,3 @@
+pydub
+SpeechRecognition
+PyYAML
